Clarify usage of `scaling`
Scaling has no effect if kind is AIRY, BAIRY, AIRYD, or BAIRYD.

Scaling=NONE should not be used with WONE, WTWO, DWONE, or DWTWO. Instead, only HUFFORD or WAIT scaling should be used for these values of `kind`.

Added checks to ensure correct usage at the beginning of the function and removed a condition which this change made irrelevant. Updated documentation.